requirements.txt
streamlit
pandas
plotly-express
streamlit-calendar
streamlit-image-comparison
gspread
gspread-dataframe
Pillow
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib